update api gateway

+++ room-management/room-domain/room-application-service/src/main/java/com/poly/room/management/domain/dto/request/CreateRoomRequest.java
@@ -18,7 +18,7 @@ public class CreateRoomRequest {
     private String roomNumber;
 
     @NotNull(message = "Room type ID is required")
-    private Long roomTypeId;
+    private Integer roomTypeId;
 
     @NotNull(message = "Floor is required")
     @Min(value = 1, message = "Floor must be at least 1")
